fixed finish
