+++ src/com/hl/hw1/Human.java
@@ -1,12 +1,22 @@
 package com.hl.hw1;
 
-import java.util.Scanner;
-
 public class Human {
 
-    private static String name;
-    private static String surname;
-    private static String patronymic;
+    static String name;
+    static String surname;
+    static String patronymic;
+
+    public static String getName() {
+        return name;
+    }
+
+    public static String getSurname() {
+        return surname;
+    }
+
+    public static String getPatronymic() {
+        return patronymic;
+    }
 
     public Human(String name, String surname) {
         this.name = name;
@@ -46,29 +56,5 @@ public class Human {
             return surname + " " + name.charAt(0) + ".";
     }
 
-    public static void main(String[] args) {
-        Scanner in = new Scanner(System.in);
 
-
-        System.out.println("Введите фамилию ");
-        Human.surname=in.nextLine();
-        System.out.println("Введите имя ");
-        Human.name=in.nextLine();
-        System.out.println("Введите отчество ");
-        Human.patronymic=in.nextLine();
-        System.out.println("Для вывода выберете: 1-полное ФИО, 2-фамилия, инициалы ");
-        String choice = in.nextLine();
-
-        switch (choice) {
-            case "1":
-                System.out.println(Human.getFullName());
-                break;
-            case "2":
-                System.out.println(Human.getShortName());
-                break;
-            default:
-                System.out.println("Неверный выбор, попробуйте снова");
-
-        }
-    }
 }
